+++ src/main/java/cn/haitaoss/tinyspringmvc/framework/handlerMapping/HandlerMapping.java
@@ -18,12 +18,4 @@ public interface HandlerMapping {
      * @return cn.haitaoss.tinyspringmvc.framework.handlerMapping.HandlerExecutionChain
      */
     HandlerExecutionChain getHandler(HttpServletRequest request) throws Exception;
-
-    /**
-     * 初始化方法
-     * @author haitao.chen
-     * email
-     * date 2021/4/24 7:10 下午
-     */
-    void init();
 }